move TraverseFilter, MonadCombine and MonadFilter (#1511)
* first try - kapt fail

* unfixed NonExistentClass error

* somehow stable tests

* fixed imports

* stable
