+++ modules/core/arrow-core-data/src/test/kotlin/arrow/core/OptionTest.kt
@@ -11,8 +11,8 @@ import arrow.core.extensions.option.monoid.monoid
 import arrow.core.extensions.option.monoidal.monoidal
 import arrow.core.extensions.option.show.show
 import arrow.core.extensions.tuple2.eq.eq
-import arrow.mtl.extensions.option.monadFilter.monadFilter
-import arrow.mtl.extensions.option.traverseFilter.traverseFilter
+import arrow.core.extensions.option.monadFilter.monadFilter
+import arrow.core.extensions.option.traverseFilter.traverseFilter
 import arrow.syntax.collections.firstOption
 import arrow.test.UnitSpec
 import arrow.test.generators.option
