+++ modules/core/arrow-core-data/src/test/kotlin/arrow/core/extensions/ComposedInstancesTest.kt
@@ -1,35 +1,35 @@
-package arrow.extensions
+package arrow.core.extensions
 
 import arrow.Kind
 import arrow.Kind2
 import arrow.core.ForFunction1
+import arrow.core.ForListK
+import arrow.core.ForNonEmptyList
 import arrow.core.ForOption
 import arrow.core.ForTuple2
 import arrow.core.Function1
+import arrow.core.ListK
+import arrow.core.NonEmptyList
 import arrow.core.Option
 import arrow.core.Some
 import arrow.core.Tuple2
-import arrow.core.fix
-import arrow.core.k
-import arrow.core.invoke
-import arrow.core.ListK
-import arrow.core.NonEmptyList
-import arrow.core.nel
 import arrow.core.extensions.function1.contravariant.contravariant
+import arrow.core.extensions.listk.applicative.applicative
+import arrow.core.extensions.listk.monoidK.monoidK
+import arrow.core.extensions.listk.semigroupK.semigroupK
+import arrow.core.extensions.nonemptylist.applicative.applicative
+import arrow.core.extensions.nonemptylist.foldable.foldable
+import arrow.core.extensions.nonemptylist.functor.functor
+import arrow.core.extensions.nonemptylist.traverse.traverse
 import arrow.core.extensions.option.applicative.applicative
 import arrow.core.extensions.option.foldable.foldable
 import arrow.core.extensions.option.functor.functor
 import arrow.core.extensions.option.traverse.traverse
 import arrow.core.extensions.tuple2.bifunctor.bifunctor
-import arrow.core.ForListK
-import arrow.core.ForNonEmptyList
-import arrow.core.extensions.nonemptylist.functor.functor
-import arrow.core.extensions.nonemptylist.applicative.applicative
-import arrow.core.extensions.nonemptylist.foldable.foldable
-import arrow.core.extensions.nonemptylist.traverse.traverse
-import arrow.core.extensions.listk.applicative.applicative
-import arrow.core.extensions.listk.semigroupK.semigroupK
-import arrow.core.extensions.listk.monoidK.monoidK
+import arrow.core.fix
+import arrow.core.invoke
+import arrow.core.k
+import arrow.core.nel
 import arrow.test.UnitSpec
 import arrow.test.laws.ApplicativeLaws
 import arrow.test.laws.BifunctorLaws
@@ -55,9 +55,9 @@ import arrow.typeclasses.NestedType
 import arrow.typeclasses.binest
 import arrow.typeclasses.biunnest
 import arrow.typeclasses.conest
+import arrow.typeclasses.counnest
 import arrow.typeclasses.nest
 import arrow.typeclasses.unnest
-import arrow.typeclasses.counnest
 import io.kotlintest.runner.junit4.KotlinTestRunner
 import org.junit.runner.RunWith
