+++ modules/core/arrow-core-data/src/test/kotlin/arrow/core/extensions/NumberHashTest.kt
@@ -1,7 +1,5 @@
-package arrow.extensions
+package arrow.core.extensions
 
-import arrow.core.extensions.eq
-import arrow.core.extensions.hash
 import arrow.test.UnitSpec
 import arrow.test.laws.HashLaws
 import io.kotlintest.runner.junit4.KotlinTestRunner
