+++ modules/core/arrow-core-data/src/test/kotlin/arrow/typeclasses/ConstTest.kt
@@ -4,7 +4,7 @@ import arrow.core.extensions.const.applicative.applicative
 import arrow.core.extensions.const.eq.eq
 import arrow.core.extensions.const.show.show
 import arrow.core.extensions.monoid
-import arrow.mtl.extensions.const.traverseFilter.traverseFilter
+import arrow.core.extensions.const.traverseFilter.traverseFilter
 import arrow.test.UnitSpec
 import arrow.test.laws.ApplicativeLaws
 import arrow.test.laws.EqLaws
