+++ modules/core/arrow-core/src/main/kotlin/arrow/core/extensions/ConstTraverseFilter.kt
@@ -1,10 +1,9 @@
-package arrow.mtl.extensions
+package arrow.core.extensions
 
 import arrow.Kind
 import arrow.core.Option
-import arrow.core.extensions.ConstTraverse
 import arrow.extension
-import arrow.mtl.typeclasses.TraverseFilter
+import arrow.typeclasses.TraverseFilter
 import arrow.typeclasses.Applicative
 import arrow.typeclasses.Const
 import arrow.typeclasses.ConstOf
