+++ modules/core/arrow-core/src/main/kotlin/arrow/core/extensions/function0.kt
@@ -1,27 +1,27 @@
 package arrow.core.extensions
 
 import arrow.Kind
-import arrow.core.Function0
+import arrow.core.Either
 import arrow.core.ForFunction0
+import arrow.core.Function0
 import arrow.core.Function0Of
 import arrow.core.extensions.function0.monad.monad
-import arrow.core.select as fun0Select
-import arrow.core.Either
-import arrow.extension
 import arrow.core.fix
 import arrow.core.invoke
 import arrow.core.k
+import arrow.extension
 import arrow.typeclasses.Applicative
 import arrow.typeclasses.Apply
 import arrow.typeclasses.Bimonad
 import arrow.typeclasses.Comonad
 import arrow.typeclasses.Functor
 import arrow.typeclasses.Monad
+import arrow.typeclasses.MonadFx
 import arrow.typeclasses.MonadSyntax
 import arrow.typeclasses.Monoid
-import arrow.typeclasses.MonadFx
 import arrow.typeclasses.Selective
 import arrow.typeclasses.Semigroup
+import arrow.core.select as fun0Select
 
 @extension
 interface Function0Semigroup<A> : Semigroup<Function0<A>> {
