+++ modules/core/arrow-core/src/main/kotlin/arrow/core/extensions/nonemptylist.kt
@@ -11,6 +11,7 @@ import arrow.core.fix
 import arrow.extension
 import arrow.typeclasses.Applicative
 import arrow.typeclasses.Apply
+import arrow.typeclasses.Bimonad
 import arrow.typeclasses.Comonad
 import arrow.typeclasses.Eq
 import arrow.typeclasses.Foldable
@@ -23,7 +24,6 @@ import arrow.typeclasses.Semigroup
 import arrow.typeclasses.SemigroupK
 import arrow.typeclasses.Show
 import arrow.typeclasses.Traverse
-import arrow.typeclasses.Bimonad
 import arrow.core.combineK as nelCombineK
 
 @extension
