+++ modules/core/arrow-core/src/main/kotlin/arrow/core/extensions/sequenceK.kt
@@ -21,6 +21,8 @@ import arrow.typeclasses.Functor
 import arrow.typeclasses.FunctorFilter
 import arrow.typeclasses.Hash
 import arrow.typeclasses.Monad
+import arrow.typeclasses.MonadCombine
+import arrow.typeclasses.MonadFilter
 import arrow.typeclasses.MonadSyntax
 import arrow.typeclasses.Monoid
 import arrow.typeclasses.MonoidK
@@ -183,6 +185,63 @@ interface SequenceKFunctorFilter : FunctorFilter<ForSequenceK> {
 
   override fun <A, B> Kind<ForSequenceK, A>.map(f: (A) -> B): SequenceK<B> =
     fix().map(f)
+}
+
+@extension
+interface SequenceKMonadFilter : MonadFilter<ForSequenceK> {
+  override fun <A> empty(): SequenceK<A> =
+    SequenceK.empty()
+
+  override fun <A, B> Kind<ForSequenceK, A>.mapFilter(f: (A) -> Option<B>): SequenceK<B> =
+    fix().mapFilter(f)
+
+  override fun <A, B> Kind<ForSequenceK, A>.ap(ff: Kind<ForSequenceK, (A) -> B>): SequenceK<B> =
+    fix().ap(ff)
+
+  override fun <A, B> Kind<ForSequenceK, A>.flatMap(f: (A) -> Kind<ForSequenceK, B>): SequenceK<B> =
+    fix().flatMap(f)
+
+  override fun <A, B> tailRecM(a: A, f: kotlin.Function1<A, SequenceKOf<Either<A, B>>>): SequenceK<B> =
+    SequenceK.tailRecM(a, f)
+
+  override fun <A, B> Kind<ForSequenceK, A>.map(f: (A) -> B): SequenceK<B> =
+    fix().map(f)
+
+  override fun <A, B, Z> Kind<ForSequenceK, A>.map2(fb: Kind<ForSequenceK, B>, f: (Tuple2<A, B>) -> Z): SequenceK<Z> =
+    fix().map2(fb, f)
+
+  override fun <A> just(a: A): SequenceK<A> =
+    SequenceK.just(a)
+}
+
+@extension
+interface SequenceKMonadCombine : MonadCombine<ForSequenceK> {
+  override fun <A> empty(): SequenceK<A> =
+    SequenceK.empty()
+
+  override fun <A, B> Kind<ForSequenceK, A>.mapFilter(f: (A) -> Option<B>): SequenceK<B> =
+    fix().mapFilter(f)
+
+  override fun <A, B> Kind<ForSequenceK, A>.ap(ff: Kind<ForSequenceK, (A) -> B>): SequenceK<B> =
+    fix().ap(ff)
+
+  override fun <A, B> Kind<ForSequenceK, A>.flatMap(f: (A) -> Kind<ForSequenceK, B>): SequenceK<B> =
+    fix().flatMap(f)
+
+  override fun <A, B> tailRecM(a: A, f: kotlin.Function1<A, SequenceKOf<Either<A, B>>>): SequenceK<B> =
+    SequenceK.tailRecM(a, f)
+
+  override fun <A, B> Kind<ForSequenceK, A>.map(f: (A) -> B): SequenceK<B> =
+    fix().map(f)
+
+  override fun <A, B, Z> Kind<ForSequenceK, A>.map2(fb: Kind<ForSequenceK, B>, f: (Tuple2<A, B>) -> Z): SequenceK<Z> =
+    fix().map2(fb, f)
+
+  override fun <A> just(a: A): SequenceK<A> =
+    SequenceK.just(a)
+
+  override fun <A> Kind<ForSequenceK, A>.combineK(y: Kind<ForSequenceK, A>): SequenceK<A> =
+    fix().sequenceCombineK(y)
 }
 
 fun <A> SequenceK.Companion.fx(c: suspend MonadSyntax<ForSequenceK>.() -> A): SequenceK<A> =
