+++ modules/core/arrow-core/src/main/kotlin/arrow/core/extensions/setk.kt
@@ -2,9 +2,9 @@ package arrow.core.extensions
 
 import arrow.Kind
 import arrow.core.Eval
-import arrow.core.Tuple2
 import arrow.core.ForSetK
 import arrow.core.SetK
+import arrow.core.Tuple2
 import arrow.core.fix
 import arrow.core.k
 import arrow.extension
@@ -18,17 +18,12 @@ import arrow.typeclasses.Semigroup
 import arrow.typeclasses.SemigroupK
 import arrow.typeclasses.Semigroupal
 import arrow.typeclasses.Show
-import kotlin.collections.emptySet
-import kotlin.collections.find
-import kotlin.collections.fold
-import kotlin.collections.map
 import arrow.core.combineK as setCombineK
-import kotlin.collections.plus as setPlus
 
 @extension
 interface SetKSemigroup<A> : Semigroup<SetK<A>> {
   override fun SetK<A>.combine(b: SetK<A>): SetK<A> =
-    (this.setPlus(b)).k()
+    (this.plus(b)).k()
 }
 
 @extension
