+++ modules/core/arrow-core/src/main/kotlin/arrow/core/extensions/tuple.kt
@@ -3,10 +3,15 @@
 package arrow.core.extensions
 
 import arrow.Kind
+import arrow.core.Either
 import arrow.core.Either.Left
 import arrow.core.Either.Right
-import arrow.extension
+import arrow.core.Eval
+import arrow.core.ForTuple2
+import arrow.core.Tuple10
 import arrow.core.Tuple2
+import arrow.core.Tuple2Of
+import arrow.core.Tuple2PartialOf
 import arrow.core.Tuple3
 import arrow.core.Tuple4
 import arrow.core.Tuple5
@@ -14,30 +19,25 @@ import arrow.core.Tuple6
 import arrow.core.Tuple7
 import arrow.core.Tuple8
 import arrow.core.Tuple9
-import arrow.core.Tuple10
-import arrow.core.Tuple2PartialOf
-import arrow.core.Tuple2Of
-import arrow.core.Either
 import arrow.core.fix
-import arrow.core.ForTuple2
-import arrow.core.toT
-import arrow.core.Eval
 import arrow.core.identity
-import arrow.typeclasses.Semigroup
-import arrow.typeclasses.Monoid
-import arrow.typeclasses.Functor
-import arrow.typeclasses.Bifunctor
-import arrow.typeclasses.Apply
+import arrow.core.toT
+import arrow.extension
 import arrow.typeclasses.Applicative
-import arrow.typeclasses.Monad
-import arrow.typeclasses.Foldable
+import arrow.typeclasses.Apply
 import arrow.typeclasses.Bifoldable
-import arrow.typeclasses.Traverse
+import arrow.typeclasses.Bifunctor
 import arrow.typeclasses.Bitraverse
-import arrow.typeclasses.Eq
-import arrow.typeclasses.Hash
-import arrow.typeclasses.Show
 import arrow.typeclasses.Comonad
+import arrow.typeclasses.Eq
+import arrow.typeclasses.Foldable
+import arrow.typeclasses.Functor
+import arrow.typeclasses.Hash
+import arrow.typeclasses.Monad
+import arrow.typeclasses.Monoid
+import arrow.typeclasses.Semigroup
+import arrow.typeclasses.Show
+import arrow.typeclasses.Traverse
 import arrow.core.extensions.traverse as tuple2Traverse
 
 // TODO this should be user driven allowing consumers to generate the tuple arities on demand to avoid cluttering arrow dependents with unused code
