+++ modules/core/arrow-core/src/main/kotlin/arrow/core/extensions/validated.kt
@@ -3,19 +3,21 @@ package arrow.core.extensions
 import arrow.Kind
 import arrow.core.Either
 import arrow.core.Eval
+import arrow.core.ForValidated
 import arrow.core.Invalid
 import arrow.core.Valid
 import arrow.core.Validated
-import arrow.core.ValidatedPartialOf
-import arrow.core.fix
-import arrow.core.ap
-import arrow.core.handleLeftWith
-import arrow.core.combineK
-import arrow.core.ForValidated
 import arrow.core.ValidatedOf
+import arrow.core.ValidatedPartialOf
+import arrow.core.ap
+import arrow.core.combineK
+import arrow.core.fix
+import arrow.core.handleLeftWith
 import arrow.extension
 import arrow.typeclasses.Applicative
 import arrow.typeclasses.ApplicativeError
+import arrow.typeclasses.Bifoldable
+import arrow.typeclasses.Bitraverse
 import arrow.typeclasses.Eq
 import arrow.typeclasses.Foldable
 import arrow.typeclasses.Functor
@@ -25,8 +27,6 @@ import arrow.typeclasses.Semigroup
 import arrow.typeclasses.SemigroupK
 import arrow.typeclasses.Show
 import arrow.typeclasses.Traverse
-import arrow.typeclasses.Bitraverse
-import arrow.typeclasses.Bifoldable
 import arrow.undocumented
 import arrow.core.traverse as validatedTraverse
