+++ modules/docs/arrow-docs/docs/docs/arrow/mtl/typeclasses/monadcombine/README.md
@@ -1,7 +1,7 @@
 ---
 layout: docs
 title: MonadCombine
-permalink: /docs/arrow/mtl/typeclasses/monadcombine/
+permalink: /docs/arrow/typeclasses/monadcombine/
 redirect_from:
   - /docs/typeclasses/monadcombine/
 ---
@@ -15,11 +15,11 @@ advanced
 
 ```kotlin:ank:replace
 import arrow.reflect.*
-import arrow.mtl.typeclasses.MonadCombine
+import arrow.typeclasses.MonadCombine
 
 TypeClass(MonadCombine::class).dtMarkdownList()
 ```
 
-ank_macro_hierarchy(arrow.mtl.typeclasses.MonadCombine)
+ank_macro_hierarchy(arrow.typeclasses.MonadCombine)
 
 TODO. Meanwhile you can find a short description in the [intro to typeclasses]({{ '/docs/typeclasses/intro/' | relative_url }}).
