+++ modules/docs/arrow-docs/docs/docs/arrow/mtl/typeclasses/traversefilter/README.md
@@ -1,7 +1,7 @@
 ---
 layout: docs
 title: TraverseFilter
-permalink: /docs/arrow/mtl/typeclasses/traversefilter/
+permalink: /docs/arrow/typeclasses/traversefilter/
 redirect_from:
   - /docs/typeclasses/traversefilter/
 ---
@@ -18,9 +18,9 @@ TODO. Meanwhile you can find a short description in the [intro to typeclasses]({
 
 ```kotlin:ank:replace
 import arrow.reflect.*
-import arrow.mtl.typeclasses.TraverseFilter
+import arrow.typeclasses.TraverseFilter
 
 TypeClass(TraverseFilter::class).dtMarkdownList()
 ```
 
-ank_macro_hierarchy(arrow.mtl.typeclasses.TraverseFilter)
+ank_macro_hierarchy(arrow.typeclasses.TraverseFilter)
