+++ modules/mtl/arrow-mtl-data/src/test/kotlin/arrow/mtl/OptionTTest.kt
@@ -2,34 +2,34 @@ package arrow.mtl
 
 import arrow.Kind
 import arrow.core.Either
+import arrow.core.ForId
+import arrow.core.ForNonEmptyList
+import arrow.core.Id
+import arrow.core.NonEmptyList
 import arrow.core.None
 import arrow.core.Option
 import arrow.core.Some
-import arrow.core.ForNonEmptyList
-import arrow.core.NonEmptyList
-import arrow.core.ForId
-import arrow.core.Id
+import arrow.core.extensions.ComposedFunctorFilter
 import arrow.core.extensions.const.divisible.divisible
 import arrow.core.extensions.id.monad.monad
 import arrow.core.extensions.monoid
+import arrow.core.extensions.nonemptylist.monad.monad
+import arrow.core.extensions.option.monad.monad
+import arrow.core.extensions.option.traverseFilter.traverseFilter
+import arrow.core.fix
+import arrow.core.value
 import arrow.fx.ForIO
 import arrow.fx.IO
 import arrow.fx.extensions.io.applicativeError.attempt
 import arrow.fx.extensions.io.async.async
-import arrow.fx.mtl.optiont.async.async
 import arrow.fx.typeclasses.seconds
-import arrow.core.extensions.nonemptylist.monad.monad
-import arrow.core.extensions.option.monad.monad
-import arrow.core.fix
-import arrow.core.value
-import arrow.mtl.extensions.ComposedFunctorFilter
+import arrow.fx.mtl.optiont.async.async
 import arrow.mtl.extensions.optiont.applicative.applicative
 import arrow.mtl.extensions.optiont.divisible.divisible
+import arrow.mtl.extensions.optiont.functorFilter.functorFilter
 import arrow.mtl.extensions.optiont.monoidK.monoidK
 import arrow.mtl.extensions.optiont.semigroupK.semigroupK
-import arrow.mtl.extensions.option.traverseFilter.traverseFilter
 import arrow.mtl.extensions.optiont.traverseFilter.traverseFilter
-import arrow.mtl.extensions.optiont.functorFilter.functorFilter
 import arrow.test.UnitSpec
 import arrow.test.laws.AsyncLaws
 import arrow.test.laws.DivisibleLaws
@@ -39,8 +39,8 @@ import arrow.test.laws.SemigroupKLaws
 import arrow.test.laws.TraverseFilterLaws
 import arrow.typeclasses.Const
 import arrow.typeclasses.Eq
-import arrow.typeclasses.NestedType
 import arrow.typeclasses.Monad
+import arrow.typeclasses.NestedType
 import arrow.typeclasses.const
 import arrow.typeclasses.nest
 import arrow.typeclasses.unnest
