+++ modules/mtl/arrow-mtl/src/main/kotlin/arrow/mtl/extensions/optiont.kt
@@ -1,52 +1,53 @@
 package arrow.mtl.extensions
 
 import arrow.Kind
-import arrow.core.ForOption
-import arrow.core.Option
 import arrow.core.Either
 import arrow.core.Eval
+import arrow.core.ForOption
 import arrow.core.None
+import arrow.core.Option
 import arrow.core.Tuple2
+import arrow.core.extensions.ComposedTraverseFilter
+import arrow.core.extensions.option.applicative.applicative
+import arrow.core.extensions.option.foldable.foldable
+import arrow.core.extensions.option.traverse.traverse
+import arrow.core.extensions.option.traverseFilter.traverseFilter
+import arrow.core.fix
+import arrow.core.identity
 import arrow.core.left
 import arrow.core.none
 import arrow.core.right
 import arrow.core.some
 import arrow.core.toT
-import arrow.core.fix
-import arrow.core.identity
-import arrow.mtl.OptionT
-import arrow.mtl.OptionTPartialOf
-import arrow.mtl.OptionTOf
-import arrow.mtl.fix
-import arrow.mtl.value
 import arrow.extension
-import arrow.core.extensions.option.applicative.applicative
-import arrow.core.extensions.option.foldable.foldable
-import arrow.core.extensions.option.traverse.traverse
-import arrow.mtl.mapFilter
-import arrow.mtl.extensions.option.traverseFilter.traverseFilter
+import arrow.mtl.OptionT
+import arrow.mtl.OptionTOf
+import arrow.mtl.OptionTPartialOf
 import arrow.mtl.extensions.optiont.monad.monad
+import arrow.mtl.fix
+import arrow.mtl.mapFilter
+import arrow.mtl.value
+import arrow.typeclasses.Applicative
 import arrow.typeclasses.ApplicativeError
+import arrow.typeclasses.ComposedTraverse
 import arrow.typeclasses.Contravariant
 import arrow.typeclasses.Decidable
 import arrow.typeclasses.Divide
 import arrow.typeclasses.Divisible
 import arrow.typeclasses.Foldable
+import arrow.typeclasses.Functor
+import arrow.typeclasses.FunctorFilter
 import arrow.typeclasses.Monad
-import arrow.typeclasses.MonadSyntax
 import arrow.typeclasses.MonadError
+import arrow.typeclasses.MonadSyntax
 import arrow.typeclasses.MonadThrow
 import arrow.typeclasses.MonoidK
-import arrow.typeclasses.ComposedTraverse
-import arrow.typeclasses.SemigroupK
-import arrow.typeclasses.FunctorFilter
-import arrow.mtl.typeclasses.TraverseFilter
-import arrow.typeclasses.Applicative
-import arrow.typeclasses.Functor
 import arrow.typeclasses.Nested
+import arrow.typeclasses.SemigroupK
 import arrow.typeclasses.Traverse
-import arrow.typeclasses.unnest
+import arrow.typeclasses.TraverseFilter
 import arrow.typeclasses.compose
+import arrow.typeclasses.unnest
 import arrow.undocumented
 
 @extension
